requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
cryptography==41.0.3
distlib==0.3.7
dj-database-url==2.1.0
Django==3.2.7
django-bootstrap4==23.2
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.1.0
django-test-without-migrations==0.6
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
filelock==3.12.4
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
httplib2==0.22.0
identify==2.5.29
idna==3.4
inflection==0.5.1
msgpack==1.0.5
nodeenv==1.8.0
packaging==23.1
phonenumbers==8.13.20
platformdirs==3.10.0
pre-commit==3.4.0
proto-plus==1.22.3
protobuf==4.24.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.5
whitenoise==6.5.0
dj-database-url==2.1.0
gunicorn==21.2.0
python-decouple==3.4
Pillow==10.0.1
django-cors-headers==3.5.0